State of Sparkly as of Feb 2022.
A code dump from my internal branch.
History can be found there.

Many many changes, mostly works. Installation is a PITA.
HA is making most automation of the installation more and more
difficult. This is an ISSUE.

Node red now used for complicated automation, that was not possible
with HA. appdaemon was an option, but NR is friendlier.
Integration is a problem, need solution for preserving NR flows on
a new install w/o having to recreate a long term token.